requirements.txt
openai==2.24.0
python-dotenv==1.2.2
ddgs==9.11.1
rich==14.3.3
python-telegram-bot==22.6
psutil==7.2.2
loguru==0.7.3
ruff==0.15.4
APScheduler==3.11.2
SQLAlchemy==2.0.48
tzlocal==5.3.1
requests==2.32.5
python-frontmatter==1.1.0